Panic: shutdown on panic
Added a new config option to shutdown the system on panic. This is
useful for testers and some systems becuse when the system panics
it sits at a high CPU usage and can cause the system to overheat and
run slow.

The kernel panic message is as follows:
```
[    0.193962] ---[ end Kernel panic - not syncing: System is deadlocked on memory ]---
[    0.194112] System powering down due to kernel panic
[    0.194966] reboot: Power down
```

The panic will run `kernel_power_off()` which should cleanly shutdown.

Signed-off-by: lewisevans2007 <60709927+lewisevans2007@users.noreply.github.com>